@@ -80,8 +80,8 @@
     "koa-router": "^7.4.0",
     "object-assign": "^4.1.0",
     "prettier": "^1.5.2",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-helmet": "^5.2.0",
     "run-verify": "^1.0.2",
     "stream-to-array": "^2.3.0",
