@@ -44,8 +44,8 @@
     "babel-register": "^6.5.2",
     "electrode-archetype-njs-module-dev": "^2.3.2",
     "electrode-server": "^1.5.1",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "run-verify": "^1.0.2",
     "stream-to-array": "^2.3.0",
     "test-init-nm": "./test/nm/test-init-nm",
@@ -60,8 +60,8 @@
     "redux": "^4.0.0 || ^3.6.0"
   },
   "peerDependencies": {
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "nyc": {
     "all": true,
