@@ -32,30 +32,29 @@
     "dist"
   ],
   "dependencies": {
-    "@babel/runtime": "^7.8.3",
+    "@babel/runtime": "^7.17.9",
     "optional-require": "^1.1.6",
     "subapp-util": "^1.1.4",
     "subapp-web": "^1.0.52"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.3",
-    "@babel/core": "^7.2.2",
-    "@babel/plugin-transform-runtime": "^7.8.3",
-    "@babel/preset-env": "^7.3.1",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/register": "^7.7.7",
+    "@babel/cli": "^7.17.10",
+    "@babel/core": "^7.17.10",
+    "@babel/plugin-transform-runtime": "^7.17.10",
+    "@babel/preset-env": "^7.17.10",
+    "@babel/preset-react": "^7.16.7",
+    "@babel/register": "^7.17.7",
     "@xarc/app": "^10.0.13",
     "babel-preset-minify": "^0.5.1",
     "electrode-archetype-njs-module-dev": "^3.0.3",
-    "jsdom": "^15.2.1",
-    "react": "^16.8.3",
-    "react-async-ssr": "^0.6.0",
-    "react-dom": "^16.8.3",
-    "react-redux": "^6.0.1",
-    "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2",
-    "redux": "^4.0.1",
-    "run-verify": "^1.2.2",
+    "jsdom": "^19.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.1",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "redux": "^4.2.0",
+    "run-verify": "^1.2.6",
     "subapp-pkg-util": "../subapp-pkg-util"
   },
   "peerDependencies": {
