@@ -42,98 +42,99 @@
     "@xarc/app": "^10.0.13"
   },
   "dependencies": {
-    "@babel/cli": "^7.1.5",
-    "@babel/core": "^7.1.6",
-    "@babel/plugin-proposal-class-properties": "^7.2.1",
-    "@babel/plugin-proposal-decorators": "^7.4.4",
-    "@babel/plugin-proposal-object-rest-spread": "^7.0.0",
-    "@babel/plugin-transform-react-constant-elements": "^7.0.0",
-    "@babel/plugin-transform-react-inline-elements": "^7.0.0",
-    "@babel/plugin-transform-runtime": "^7.1.0",
-    "@babel/preset-env": "^7.1.6",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.0.0",
-    "@babel/register": "^7.0.0",
-    "@jchip/redbird": "^1.2.2",
+    "@babel/cli": "^7.17.10",
+    "@babel/core": "^7.17.12",
+    "@babel/plugin-proposal-class-properties": "^7.17.12",
+    "@babel/plugin-proposal-decorators": "^7.17.12",
+    "@babel/plugin-proposal-object-rest-spread": "^7.17.12",
+    "@babel/plugin-transform-react-constant-elements": "^7.17.12",
+    "@babel/plugin-transform-react-inline-elements": "^7.16.7",
+    "@babel/plugin-transform-runtime": "^7.17.12",
+    "@babel/preset-env": "^7.17.12",
+    "@babel/preset-react": "^7.17.12",
+    "@babel/preset-typescript": "^7.17.12",
+    "@babel/register": "^7.17.7",
+    "@jchip/redbird": "^1.3.0",
     "@xarc/dev-base": "^0.1.0",
     "@xarc/run": "^1.0.5",
     "@xarc/subapp": "^0.3.3",
     "@xarc/webpack": "^11.1.1",
-    "ansi-to-html": "^0.6.8",
+    "ansi-to-html": "^0.7.2",
     "babel-plugin-lodash": "^3.3.4",
-    "babel-plugin-minify-dead-code-elimination": "^0.5.0",
+    "babel-plugin-minify-dead-code-elimination": "^0.5.2",
     "babel-plugin-react-css-modules": "^5.2.6",
     "babel-plugin-transform-node-env-inline": "^0.4.3",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.20",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.24",
     "boxen": "^5.0.0",
     "chalker": "^1.2.0",
-    "chokidar": "^3.4.1",
-    "core-js": "^3.6.5",
+    "chokidar": "^3.5.3",
+    "core-js": "^3.22.5",
     "electrode-node-resolver": "^2.0.2",
     "fast-async": "^7.0.5",
-    "filter-scan-dir": "1.1.1",
-    "fs-extra": "^9.1.0",
+    "filter-scan-dir": "1.5.5",
+    "fs-extra": "^10.1.0",
     "identity-obj-proxy": "^3.0.0",
     "is-ci": "^2.0.0",
-    "lodash": "^4.17.19",
-    "log-update": "^4.0.0",
-    "mime": "^2.4.6",
+    "lodash": "^4.17.21",
+    "log-update": "^5.0.1",
+    "mime": "^3.0.0",
     "mkdirp": "^1.0.4",
-    "nix-clap": "^1.3.12",
+    "nix-clap": "^1.3.13",
     "nyc": "^15.1.0",
-    "optional-require": "^1.1.6",
+    "optional-require": "^1.1.8",
     "pkg-up": "^3.1.0",
-    "prompts": "^2.3.2",
+    "prompts": "^2.4.2",
     "ps-get": "^1.1.0",
     "read-pkg-up": "^7.0.1",
-    "regenerator-runtime": "^0.13.7",
+    "regenerator-runtime": "^0.13.9",
     "request": "^2.88.2",
     "require-at": "^1.0.6",
-    "semver": "^7.3.2",
+    "semver": "^7.3.7",
     "serve-index-fs": "^1.10.1",
     "subapp-util": "^1.1.4",
-    "sudo-prompt": "^8.2.5",
-    "tslib": "^2.1.0",
+    "sudo-prompt": "^9.2.1",
+    "tslib": "^2.4.0",
     "visual-logger": "^1.1.3",
     "webpack-dev-middleware": "^4.1.0",
-    "webpack-hot-middleware": "^2.25.0",
-    "winston": "^2.4.4",
-    "xaa": "^1.7.0",
+    "webpack-hot-middleware": "^2.25.1",
+    "winston": "^3.7.2",
+    "xaa": "^1.7.3",
     "xenv-config": "^1.3.1",
     "xsh": "^0.4.5"
   },
   "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@types/chai": "^4.2.14",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@types/chai": "^4.3.1",
     "@types/mocha": "^8.2.0",
     "@types/node": "^14.14.16",
     "@types/sinon": "^9.0.10",
-    "@types/sinon-chai": "^3.2.5",
+    "@types/sinon-chai": "^3.2.8",
     "@typescript-eslint/eslint-plugin": "^4.11.0",
     "@typescript-eslint/parser": "^4.11.0",
     "@xarc/app": "^10.0.13",
-    "@xarc/module-dev": "^3.0.2",
-    "@xarc/run": "^1.0.4",
+    "@xarc/module-dev": "^4.1.0",
+    "@xarc/run": "^1.1.1",
     "babel-eslint": "^10.1.0",
-    "chai": "^4.2.0",
+    "chai": "^4.3.6",
     "e1": "./e1",
-    "electrode-server": "^3.2.0",
+    "electrode-server": "^3.3.0",
     "eslint": "^7.16.0",
     "eslint-config-walmart": "^2.2.1",
-    "eslint-plugin-filenames": "^1.1.0",
+    "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-jsdoc": "^32.0.2",
+    "eslint-plugin-tsdoc": "^0.2.16",
     "mocha": "^8.2.1",
     "mock-require": "^3.0.3",
     "nyc": "^15.1.0",
-    "prettier": "^2.2.1",
-    "run-verify": "^1.2.5",
-    "shx": "^0.3.2",
+    "prettier": "^2.6.2",
+    "run-verify": "^1.2.6",
+    "shx": "^0.3.4",
     "sinon": "^9.2.2",
-    "sinon-chai": "^3.5.0",
-    "source-map-support": "^0.5.19",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
     "ts-node": "^10.7.0",
-    "typedoc": "^0.20.13",
-    "typescript": "^4.1.3"
+    "typedoc": "^0.22.15",
+    "typescript": "^4.6.4"
   },
   "engines": {
     "node": ">= 12",
