@@ -32,40 +32,41 @@
     "Joel Chen <xchen@walmartlabs.com>"
   ],
   "dependencies": {
-    "@babel/runtime": "^7.12.5",
+    "@babel/runtime": "^7.17.9",
     "isomorphic-loader": "^4.5.0",
-    "optional-require": "^1.1.6",
+    "optional-require": "^1.1.8",
     "subapp-util": "^1.1.4",
-    "tslib": "^2.1.0"
+    "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^1.0.1",
-    "@types/chai": "^4.2.14",
-    "@types/mocha": "^8.2.0",
-    "@types/node": "^14.14.16",
-    "@types/sinon": "^9.0.10",
-    "@types/sinon-chai": "^3.2.5",
-    "@typescript-eslint/eslint-plugin": "^4.11.0",
-    "@typescript-eslint/parser": "^4.11.0",
-    "@xarc/module-dev": "^3.0.2",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
+    "@types/chai": "^4.3.1",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^17.0.31",
+    "@types/sinon": "^10.0.11",
+    "@types/sinon-chai": "^3.2.8",
+    "@typescript-eslint/eslint-plugin": "^5.22.0",
+    "@typescript-eslint/parser": "^5.22.0",
+    "@xarc/module-dev": "^4.1.0",
     "babel-eslint": "^10.1.0",
-    "chai": "^4.2.0",
-    "chalk": "^1.1.3",
-    "eslint": "^7.16.0",
+    "chai": "^4.3.6",
+    "chalk": "^5.0.1",
+    "eslint": "^8.14.0",
     "eslint-config-walmart": "^2.2.1",
-    "eslint-plugin-filenames": "^1.1.0",
+    "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-jsdoc": "^30.7.9",
-    "mocha": "^8.2.1",
+    "eslint-plugin-tsdoc": "^0.2.16",
+    "mocha": "^10.0.0",
     "mock-require": "^3.0.3",
     "nyc": "^15.1.0",
-    "prettier": "^1.14.2",
-    "prompts": "^2.2.1",
-    "sinon": "^9.2.2",
-    "sinon-chai": "^3.5.0",
-    "source-map-support": "^0.5.19",
-    "ts-node": "^9.1.1",
-    "typedoc": "^0.20.13",
-    "typescript": "^4.1.3",
+    "prettier": "^2.6.2",
+    "prompts": "^2.4.2",
+    "sinon": "^13.0.2",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
+    "ts-node": "^10.7.0",
+    "typedoc": "^0.22.15",
+    "typescript": "^4.6.4",
     "xsh": "^0.4.5"
   },
   "engines": {
