@@ -29,40 +29,41 @@
   },
   "dependencies": {
     "@xarc/subapp": "^0.3.3",
-    "react-query": "~3.29.0",
-    "tslib": "^2.1.0"
+    "react-query": "~3.39.0",
+    "tslib": "^2.4.0"
   },
   "devDependencies": {
-    "@istanbuljs/nyc-config-typescript": "^1.0.1",
+    "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@testing-library/react": "^11.2.3",
-    "@types/chai": "^4.2.14",
-    "@types/mocha": "^8.2.0",
-    "@types/node": "^14.14.16",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/sinon": "^9.0.10",
-    "@types/sinon-chai": "^3.2.5",
-    "@typescript-eslint/eslint-plugin": "^4.11.0",
-    "@typescript-eslint/parser": "^4.11.0",
-    "@xarc/module-dev": "^3.0.2",
+    "@types/chai": "^4.3.1",
+    "@types/mocha": "^9.1.1",
+    "@types/node": "^18.0.0",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
+    "@types/sinon": "^10.0.11",
+    "@types/sinon-chai": "^3.2.8",
+    "@typescript-eslint/eslint-plugin": "^5.29.0",
+    "@typescript-eslint/parser": "^5.29.0",
+    "@xarc/module-dev": "^4.1.0",
     "babel-eslint": "^10.1.0",
-    "chai": "^4.2.0",
-    "eslint": "^7.16.0",
+    "chai": "^4.3.6",
+    "eslint": "^8.18.0",
     "eslint-config-walmart": "^2.2.1",
-    "eslint-plugin-filenames": "^1.1.0",
+    "eslint-plugin-filenames": "^1.3.2",
     "eslint-plugin-jsdoc": "^30.7.9",
+    "eslint-plugin-tsdoc": "^0.2.16",
     "jsdom": "^16.4.0",
     "jsdom-global": "^3.0.2",
-    "mocha": "^8.2.1",
+    "mocha": "^10.0.0",
     "nyc": "^15.1.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "sinon": "^9.2.2",
-    "sinon-chai": "^3.5.0",
-    "source-map-support": "^0.5.19",
-    "ts-node": "^9.1.1",
-    "typedoc": "^0.20.13",
-    "typescript": "^4.1.3"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "sinon": "^14.0.0",
+    "sinon-chai": "^3.7.0",
+    "source-map-support": "^0.5.21",
+    "ts-node": "^10.8.1",
+    "typedoc": "^0.22.17",
+    "typescript": "^4.7.4"
   },
   "files": [
     "dist",
