@@ -56,8 +56,8 @@
     "jsdom-global": "^3.0.2",
     "mocha": "^8.2.1",
     "nyc": "^15.1.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sinon": "^9.2.2",
     "sinon-chai": "^3.5.0",
     "source-map-support": "^0.5.19",
