@@ -1,13 +1,21 @@
+/* eslint-disable */
 "use strict";
 
 const url = require("url");
 const React = require("react"); // eslint-disable-line
 const lib = require("../../lib");
-const { withRouter } = require("react-router");
-const { Route, Switch } = require("react-router-dom"); // eslint-disable-line
+const { Routes, Route } = require("react-router-dom"); // eslint-disable-line
 const { asyncVerify } = require("run-verify");
 const Redux = require("redux");
 const { connect } = require("react-redux");
+
+const getStreamWritable = (stream, next) => {
+  let res = "";
+  stream.on("data", data => (res += data.toString()));
+  stream.on("end", () => next(null, res));
+  stream.on("error", next);
+  return res;
+};
 
 describe("SSR React framework", function () {
   it("should setup React framework", () => {
@@ -37,7 +45,7 @@ describe("SSR React framework", function () {
     expect(res).equals("");
   });
 
-  it("should render subapp with w/o initial props if no prepare provided", async () => {
+  it("should render subapp with w/o initial props if no prepare provided", () => {
     const framework = new lib.FrameworkLib({
       subApp: {
         Component: props => {
@@ -50,11 +58,15 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello <");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
+      res => expect(res).contains(`<div>Hello </div>`)
+    );
   });
 
-  it("should render Component from subapp with initial props from prepare", async () => {
+  it("should render Component from subapp with initial props from prepare", () => {
     const framework = new lib.FrameworkLib({
       subApp: {
         prepare: () => ({ test: "foo bar" }),
@@ -68,8 +80,12 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello foo bar");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
+      res => expect(res).contains(`<div>Hello <!-- -->foo bar</div>`)
+    );
   });
 
   it("should allow preparing data before SSR", async () => {
@@ -88,8 +104,12 @@ describe("SSR React framework", function () {
     });
     await framework.handlePrepare();
     expect(framework._initialProps).to.be.ok;
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello foo bar");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
+      res => expect(res).contains(`<div>Hello <!-- -->foo bar</div>`)
+    );
   });
 
   it("should render Component with stream if enabled", () => {
@@ -108,69 +128,13 @@ describe("SSR React framework", function () {
     });
     return asyncVerify(
       () => framework.handleSSR(),
-      (stream, next) => {
-        let res = "";
-        stream.on("data", data => (res += data.toString()));
-        stream.on("end", () => next(null, res));
-        stream.on("error", next);
-      },
-      res => expect(res).contains("Hello foo bar")
-    );
-  });
-
-  it("should hydrate render Component with stream if enabled", () => {
-    const framework = new lib.FrameworkLib({
-      subApp: {
-        prepare: () => ({ test: "foo bar" }),
-        Component: props => {
-          return <div>Hello {props.test}</div>;
-        }
-      },
-      subAppServer: {},
-      options: { serverSideRendering: true, useStream: true, hydrateServerData: true },
-      context: {
-        user: {}
-      }
-    });
-    return asyncVerify(
-      () => framework.handleSSR(),
-      (stream, next) => {
-        let res = "";
-        stream.on("data", data => (res += data.toString()));
-        stream.on("end", () => next(null, res));
-        stream.on("error", next);
-      },
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
       res => expect(res).contains(`<div>Hello <!-- -->foo bar</div>`)
     );
   });
 
-  it("should render Component from subapp with hydration info", async () => {
-    const framework = new lib.FrameworkLib({
-      subApp: {
-        prepare: () => ({
-          test: "foo bar"
-        }),
-        Component: props => {
-          return <div>Hello {props.test}</div>;
-        }
-      },
-      subAppServer: {},
-      options: {
-        serverSideRendering: true,
-        hydrateServerData: true
-      },
-      context: {
-        user: {}
-      }
-    });
-    // data-reactroot isn't getting created due to Context.Provider
-    // see https://github.com/facebook/react/issues/15012
-    const res = await framework.handleSSR();
-    // but the non-static renderToString adds a <!-- --> for some reason
-    expect(res).contains("Hello <!-- -->foo bar");
-  });
-
-  it("should render Component from subapp with initial props from server's prepare", async () => {
+  it("should render Component from subapp with initial props from server's prepare", () => {
     const framework = new lib.FrameworkLib({
       subApp: {
         Component: props => {
@@ -185,11 +149,15 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello foo bar");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
+      res => expect(res).contains(`<div>Hello <!-- -->foo bar</div>`)
+    );
   });
 
-  it("should render Component from subapp with initial props from server's prepare while using attachInitialState", async () => {
+  it("should render Component from subapp with initial props from server's prepare while using attachInitialState", () => {
     const framework = new lib.FrameworkLib({
       subApp: {
         Component: props => {
@@ -205,8 +173,12 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello foo bar");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
+      res => expect(res).contains(`<div>Hello <!-- -->foo bar</div>`)
+    );
   });
 
   it("should init redux store in context and render Component", async () => {
@@ -237,11 +209,22 @@ describe("SSR React framework", function () {
         options: { serverSideRendering: true },
         context
       });
-      const res = await framework.handleSSR();
-      expect(res).contains("Hello foo bar");
-      expect(framework.initialStateStr).equals(`{"test":"foo bar"}`);
-      expect(context.user).to.have.property("storeContainer");
-      expect(storeReady).equal(true);
+
+      return asyncVerify(
+        () => framework.handleSSR(),
+        (stream, next) => {
+          let res = "";
+          stream.on("data", data => (res += data.toString()));
+          stream.on("end", () => next(null, res));
+          stream.on("error", next);
+        },
+        res => {
+          expect(res).contains(`<div>Hello <!-- -->foo bar</div>`);
+          expect(framework.initialStateStr).equals(`{"test":"foo bar"}`);
+          expect(context.user).to.have.property("storeContainer");
+          expect(storeReady).equal(true);
+        }
+      );
     };
     await verify();
     const store = context.user.storeContainer.store;
@@ -250,7 +233,7 @@ describe("SSR React framework", function () {
     expect(store).to.equal(context.user.storeContainer.store);
   });
 
-  it("should init redux store and render Component but doesn't attach initial state", async () => {
+  it("should init redux store and render Component but doesn't attach initial state", () => {
     const Component = connect(x => x)(props => <div>Hello {props.test}</div>);
 
     const framework = new lib.FrameworkLib({
@@ -266,9 +249,15 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello foo bar");
-    expect(framework.initialStateStr).equals(undefined);
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      res => {
+        // renderToPipeableSrteam adds a <!-- --> before props
+        expect(res).contains(`<div>Hello <!-- -->foo bar</div>`);
+        expect(framework.initialStateStr).equals(undefined);
+      }
+    );
   });
 
   it("should init redux store but doesn't render Component if serverSideRendering is not true", async () => {
@@ -292,7 +281,7 @@ describe("SSR React framework", function () {
     expect(framework.initialStateStr).equals(undefined);
   });
 
-  it("should init redux store with empty state without prepare and render Component", async () => {
+  it("should init redux store with empty state without prepare and render Component", () => {
     const Component = connect(x => x)(props => <div>Hello {props.test}</div>);
 
     const framework = new lib.FrameworkLib({
@@ -307,36 +296,17 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello <");
-    expect(framework.initialStateStr).equals(`{}`);
-  });
-
-  it("should hydrate render Component with suspense using react-async-ssr", async () => {
-    const framework = new lib.FrameworkLib({
-      subApp: {
-        Component: props => {
-          return (
-            <React.Suspense fallback={<h1>Loading...</h1>}>
-              <div>Hello {props.test}</div>
-            </React.Suspense>
-          );
-        }
-      },
-      subAppServer: {
-        prepare: () => ({ test: "foo bar" })
-      },
-      options: { serverSideRendering: true, suspenseSsr: true, hydrateServerData: true },
-      context: {
-        user: {}
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      res => {
+        expect(res).contains("Hello <");
+        expect(framework.initialStateStr).equals(`{}`);
       }
-    });
-    const res = await framework.handleSSR();
-    // react-async-ssr includes data-reactroot
-    expect(res).contains(`<div data-reactroot="">Hello <!-- -->foo bar</div>`);
+    );
   });
 
-  it("should render Component with suspense using react-async-ssr", async () => {
+  it("should render Component with suspense using react-async-ssr", () => {
     const framework = new lib.FrameworkLib({
       subApp: {
         Component: props => {
@@ -355,11 +325,15 @@ describe("SSR React framework", function () {
         user: {}
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello foo bar");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      // renderToPipeableSrteam adds a <!-- --> before props
+      res => expect(res).contains(`<div>Hello <!-- -->foo bar</div>`)
+    );
   });
 
-  it("should render Component with react context containing request", async () => {
+  it("should render Component with react context containing request", () => {
     const request = {};
     const framework = new lib.FrameworkLib({
       subApp: {
@@ -386,23 +360,31 @@ describe("SSR React framework", function () {
         user: { request }
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains(`<div>IS_SSR: true HAS_REQUEST: yes</div>`);
-    expect(request.foo).equals("bar");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      res => {
+        // renderToPipeableSrteam adds a <!-- --> before props
+        expect(res).contains(`<div>IS_SSR: <!-- -->true<!-- --> HAS_REQUEST: <!-- -->yes</div>`);
+        expect(request.foo).equals("bar");
+      }
+    );
+
   });
 
   it("should render subapp with react-router StaticRouter", async () => {
-    const TestComponent = () => {
-      return <div>Hello test path</div>;
-    };
-    const Component = withRouter(props => {
-      return (
-        <Switch>
-          <Route path="/test" component={TestComponent} {...props} />
-          <Route path="/foo" component={() => "bar"} {...props} />
-        </Switch>
-      );
-    });
+    /* eslint-disable no-unused-vars */
+    const TestComponent = () => <div>Hello test path</div>;
+    const FooBar = () => <div>foo</div>;
+    /* eslint-enable no-unused-vars */
+
+    const Component = () => (
+      <Routes>
+        <Route path="/test" element={<TestComponent />} />
+        <Route path="/foo" element={<FooBar />} />
+      </Routes>
+    );
+
     const framework = new lib.FrameworkLib({
       subApp: {
         useReactRouter: true,
@@ -418,15 +400,17 @@ describe("SSR React framework", function () {
         }
       }
     });
-    const res = await framework.handleSSR();
-    expect(res).contains("Hello test path<");
+    return asyncVerify(
+      () => framework.handleSSR(),
+      (stream, next) => getStreamWritable(stream, next),
+      res => expect(res).contains("Hello test path<")
+    );
   });
 
   it("should render subapp with custom renderer e.g. css in js solution", async () => {
     const Component = () => {
       return <div>Hello test path</div>;
     };
-
     const framework = new lib.FrameworkLib({
       subApp: {
         Component
