@@ -4,7 +4,6 @@ const ReduxRouterEngine = require("../..");
 const expect = require("chai").expect;
 const electrodeServer = require("electrode-server");
 const { runFinally, asyncVerify } = require("run-verify");
-const streamToArray = require("stream-to-array");
 
 require("babel-register");
 
@@ -25,15 +24,7 @@ describe("electrode server (Hapi) integration", function() {
       handler: async (request, reply) => {
         if (!engine) engine = new ReduxRouterEngine({ routes, streaming, withIds });
         const result = await engine.render(request);
-        if (streaming) {
-          result.resultType = result.html.constructor.name;
-          streamToArray(result.html, (err, arr) => {
-            result.html = arr.join("");
-            reply(result);
-          });
-        } else {
-          reply(result);
-        }
+        reply(result);
       }
     });
 
@@ -55,7 +46,7 @@ describe("electrode server (Hapi) integration", function() {
       () => setupServer(),
       s => {
         server = s;
-        return server.inject("/test").then(resp => {
+        return server.inject("/test").then((resp) => {
           expect(resp.result).to.deep.equal({
             status: 200,
             html: "<div>Page<div>Home</div></div>",
@@ -73,10 +64,10 @@ describe("electrode server (Hapi) integration", function() {
       () => setupServer(),
       s => {
         server = s;
-        return server.inject("/test?foo=bar").then(resp => {
+        return server.inject("/test?foo=bar").then((resp) => {
           expect(resp.result).to.deep.equal({
             status: 200,
-            html: "<div>Page<div>Home - Query: ?foo=bar</div></div>",
+            html: "<div>Page<div>Home<!-- --> - Query: ?foo=bar</div></div>",
             prefetch: "window.__PRELOADED_STATE__ = {};"
           });
         });
@@ -91,12 +82,11 @@ describe("electrode server (Hapi) integration", function() {
       () => setupServer(true),
       s => {
         server = s;
-        return server.inject("/test").then(resp => {
+        return server.inject("/test").then((resp) => {
           expect(resp.result).to.deep.equal({
             status: 200,
             html: "<div>Page<div>Home</div></div>",
-            prefetch: "window.__PRELOADED_STATE__ = {};",
-            resultType: "ReactMarkupReadableStream"
+            prefetch: "window.__PRELOADED_STATE__ = {};"
           });
         });
       },
@@ -110,12 +100,11 @@ describe("electrode server (Hapi) integration", function() {
       () => setupServer(true, true),
       s => {
         server = s;
-        return server.inject("/test").then(resp => {
+        return server.inject("/test").then((resp) => {
           expect(resp.result).to.deep.equal({
             status: 200,
             html: `<div>Page<div>Home</div></div>`,
-            prefetch: "window.__PRELOADED_STATE__ = {};",
-            resultType: "ReactMarkupReadableStream"
+            prefetch: "window.__PRELOADED_STATE__ = {};"
           });
         });
       },
