@@ -1,5 +1,5 @@
 import { SubAppDef, SubAppFeatureFactory } from "@xarc/subapp";
-import { Router, BrowserRouter } from "react-router-dom";
+import { BrowserRouter, unstable_HistoryRouter as HistoryRouter } from "react-router-dom";
 import { ReactRouterFeatureOptions, _id, _subId } from "../common";
 import { createBrowserHistory } from "history";
 
@@ -24,20 +24,6 @@ export function reactRouterFeature(options: ReactRouterFeatureOptions): SubAppFe
       let history: any;
       let TheRouter: any;
 
-      if (!options.hasOwnProperty("history") || options.history === true) {
-        // if user didn't specify history or set it to true, then create a shared
-        // history and use Router
-        history = staticHistory || (staticHistory = createBrowserHistory());
-        TheRouter = Router;
-      } else if (options.history) {
-        // use Router with the history user provided
-        history = options.history;
-        TheRouter = Router;
-      } else {
-        // user specify a falsy history, fallback to BrowserRouter
-        TheRouter = BrowserRouter;
-      }
-
       subapp._features.reactRouter = {
         id,
         subId,
@@ -45,8 +31,34 @@ export function reactRouterFeature(options: ReactRouterFeatureOptions): SubAppFe
           const Component = input.Component || subapp._getExport()?.Component;
           return {
             Component: (props: any) => {
+              /**
+               * unstable_HistoryRouter as HistoryRouter is used here instead of <Router>
+               * because react_router v6 removed history prop support for <Router>.
+               * This API is currently prefixed as unstable_ because you may unintentionally
+               * add two versions of the history library, the one you have added
+               * to the package.json and whatever version React Router uses internally.
+               * Once react-router has a mechanism to detect mis-matched versions,
+               * this API will remove its unstable_ prefix.
+               * https://reactrouter.com/docs/en/v6/routers/history-router
+               * For now, use the unstable_HistoryRouter version here.
+               */
+              if (!options.hasOwnProperty("history") || options.history === true) {
+                // if user didn't specify history or set it to true, then create a shared
+                // history and use HistoryRouter
+                history = staticHistory || (staticHistory = createBrowserHistory());
+                TheRouter = HistoryRouter;
+              } else if (options.history) {
+                // use HistoryRouter with the history user provided
+                history = options.history;
+                TheRouter = HistoryRouter;
+              } else {
+                // user specify a falsy history, fallback to BrowserRouter
+                TheRouter = BrowserRouter;
+              }
+              // conditional history props. BrowserRouter does not support history props in v6.
+              const historyProps = history ? { history } : {};
               return (
-                <TheRouter history={history}>
+                <TheRouter {...historyProps} >
                   <Component {...props} />
                 </TheRouter>
               );
