@@ -1,5 +1,5 @@
 import { SubAppDef, SubAppFeatureFactory } from "@xarc/subapp";
-import { StaticRouter } from "react-router-dom";
+import { StaticRouter } from "react-router-dom/server";
 import { ReactRouterFeatureOptions, _id, _subId } from "../common";
 
 /**
@@ -10,7 +10,6 @@ import { ReactRouterFeatureOptions, _id, _subId } from "../common";
  */
 export function reactRouterFeature(options: ReactRouterFeatureOptions): SubAppFeatureFactory {
   const { createElement } = options.React; // eslint-disable-line
-
   const id = _id;
   const subId = _subId;
   return {
@@ -22,13 +21,10 @@ export function reactRouterFeature(options: ReactRouterFeatureOptions): SubAppFe
         subId,
         execute({ input, ssrData }) {
           const Component = input.Component || subapp._getExport()?.Component;
-          const routerContext = {};
-          ssrData.context.user.routerContext = routerContext;
-
           return {
             Component: () => {
               return (
-                <StaticRouter location={ssrData.path} context={routerContext}>
+                <StaticRouter location={ssrData.path}>
                   <Component />
                 </StaticRouter>
               );
